backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pandas
numpy
scipy
python-dotenv
pydantic
